fix(sqlgen,cdc,federated): project EAV-only integer/smallint by storage width (BIGINT) on every DuckDB tier (#384)
eav_data.value_numeric is unconstrained NUMERIC, so casting the DuckDB
projection by declared width TRY_CAST a stored out-of-range value to NULL
on every DuckDB leg while hot Postgres matched it. All four projection
surfaces move in lockstep: the hot EAV pivot (buildEAVPivotExpr), list
element casts (eavElementCastExpr), the CDC parquet export (castEAVValue),
and the cold NULL-augmentation type (DuckDBNullScanType, now
binding-aware). Column-bound integer/smallint keep declared width: their
physical storage is int4/int2. Pre-change parquet files carrying INT32
columns promote losslessly to BIGINT via the existing union_by_name scans
(read + compaction verified).

+++ internal/federated/cold_columns_test.go
@@ -22,7 +22,8 @@ func TestColdMissingColumnsDetectsAbsentAttributes(t *testing.T) {
 	}
 	got := coldMissingColumns(cache, union)
 	require.Equal(t, []sqlgen.NullScanColumn{
-		{Name: "score", DuckDBType: "INTEGER"},
+		// EAV-only integer augments at storage width BIGINT (#384).
+		{Name: "score", DuckDBType: "BIGINT"},
 		{Name: "tags", DuckDBType: "BIGINT[]"},
 	}, got, "absent attrs detected, present ones skipped, sorted by name")
 }
